split fake service replay and dispatch tests
