+++ docs/internal/baselines/backend-package-file-count.json
@@ -123,7 +123,7 @@
     },
     {
       "packagePath": "pkg/services/factory_sessions/internal/execution",
-      "fileCount": 31,
+      "fileCount": 32,
       "owner": "factory-session-maintainers",
       "removalReason": "Remove this ratchet after the remaining execution test seams are split into focused files."
     },
