+++ pkg/services/factory_sessions/internal/execution/fake_service_replay_test.go
@@ -1,19 +1,14 @@
-// backendsizecheck:ignore-file consolidated fake-service contract and projection tests remain together until dedicated execution test seams split.
-// pkgmaintcheck:ignore-file-lines consolidated fake-service contract and projection tests remain together until dedicated execution test seams split.
 package factorysessionexecution
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
-	"os"
 	"reflect"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/fileeffects"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 )
@@ -799,578 +794,5 @@ func TestReplaySessionProjection_EquivalentOrchestratorsPreserveResultAvailabili
 				}
 			}
 		})
-	}
-}
-
-func TestAppendDispatchInterruptedEvent_RecordsCanonicalMetadata(t *testing.T) {
-	t.Parallel()
-	startedAt := time.Date(2026, 6, 20, 12, 0, 0, 0, time.UTC)
-	session := SessionReadResult{
-		SessionID:        "dur-sess-interrupt-001",
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: "JAVASCRIPT",
-		Dialect:          "you-workflow-v1",
-		Phase:            "execute",
-		Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt},
-	}
-	base := BuildCanonicalRuntimeSessionEvents(session, ResultReadResult{
-		SessionID:    session.SessionID,
-		ResultStatus: ResultStatusNotReady,
-	})
-	dispatch := DispatchSummary{
-		ID:     "disp-js-002",
-		Status: DispatchStatusRunning,
-		Phase:  "execute",
-		Label:  "audit",
-	}
-	events := AppendDispatchInterruptedEvent(
-		base,
-		session,
-		dispatch,
-		InterruptDispatchRequest{
-			ControlRequest: ControlRequest{Reason: "operator stop"},
-			DispatchID:     "disp-js-002",
-		},
-		DispatchStatusRunning,
-		canonicalEventSourceRuntimeService,
-		startedAt,
-	)
-	if len(events) != len(base)+1 {
-		t.Fatalf("events = %d, want %d", len(events), len(base)+1)
-	}
-
-	var envelope canonicalFactoryEvent
-	if err := json.Unmarshal(events[len(events)-1], &envelope); err != nil {
-		t.Fatalf("unmarshal interrupted event: %v", err)
-	}
-	if envelope.Type != "DISPATCH_INTERRUPTED" {
-		t.Fatalf("type = %q, want DISPATCH_INTERRUPTED", envelope.Type)
-	}
-	if envelope.Context.DispatchID == nil || *envelope.Context.DispatchID != "disp-js-002" {
-		t.Fatalf("dispatchId = %#v, want disp-js-002", envelope.Context.DispatchID)
-	}
-
-	var payload dispatchInterruptedEventPayload
-	if err := json.Unmarshal(envelope.Payload, &payload); err != nil {
-		t.Fatalf("unmarshal payload: %v", err)
-	}
-	if payload.Reason != "operator stop" {
-		t.Fatalf("reason = %q, want operator stop", payload.Reason)
-	}
-	if payload.ObservedStatus != string(factoryapi.FactoryDispatchStatusRUNNING) {
-		t.Fatalf("observedStatus = %q, want RUNNING", payload.ObservedStatus)
-	}
-	if payload.RetryPlanned {
-		t.Fatal("retryPlanned = true, want false")
-	}
-}
-
-func TestMarkDispatchInterrupted_UpdatesInspectionProjection(t *testing.T) {
-	t.Parallel()
-	dispatches := []DispatchSummary{{
-		ID:     "disp-js-002",
-		Status: DispatchStatusRunning,
-	}}
-	updated, _ := MarkDispatchInterrupted(
-		dispatches,
-		map[string][]DispatchStatus{},
-		"disp-js-002",
-		InterruptDispatchRequest{DispatchID: "disp-js-002"},
-	)
-	if updated[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("status = %q, want INTERRUPTED", updated[0].Status)
-	}
-	if updated[0].FailureDetail == nil || updated[0].FailureDetail.Reason != dispatchInterruptionFailureReasonCode {
-		t.Fatalf("failureDetail = %#v, want DISPATCH_INTERRUPTED reason", updated[0].FailureDetail)
-	}
-	if updated[0].FailureDetail.Message != defaultDispatchInterruptionReason {
-		t.Fatalf("failure message = %q", updated[0].FailureDetail.Message)
-	}
-}
-
-func TestReplayDispatchProjection_DerivesInterruptedDispatchMetadata(t *testing.T) {
-	t.Parallel()
-	startedAt := time.Date(2026, 6, 20, 12, 0, 0, 0, time.UTC)
-	session := SessionReadResult{
-		SessionID:        "dur-sess-interrupt-002",
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: "JAVASCRIPT",
-		Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt},
-	}
-	events := AppendDispatchInterruptedEvent(
-		BuildCanonicalRuntimeSessionEvents(session, ResultReadResult{SessionID: session.SessionID}),
-		session,
-		DispatchSummary{ID: "disp-js-002", Status: DispatchStatusRunning, Phase: "execute"},
-		InterruptDispatchRequest{
-			ControlRequest: ControlRequest{Reason: "bad prompt"},
-			DispatchID:     "disp-js-002",
-		},
-		DispatchStatusRunning,
-		canonicalEventSourceRuntimeService,
-		startedAt,
-	)
-
-	replayed, err := ReplayDispatchProjection(events)
-	if err != nil {
-		t.Fatalf("ReplayDispatchProjection: %v", err)
-	}
-	if len(replayed) != 1 {
-		t.Fatalf("replayed dispatches = %#v, want one interrupted dispatch", replayed)
-	}
-	if replayed[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("status = %q, want INTERRUPTED", replayed[0].Status)
-	}
-	if replayed[0].FailureDetail == nil || replayed[0].FailureDetail.Message != "bad prompt" {
-		t.Fatalf("failureDetail = %#v, want bad prompt", replayed[0].FailureDetail)
-	}
-}
-
-// pkgmaintcheck:ignore-cyclomatic-complexity this fake-service regression keeps interrupt event payload assertions together on one scenario.
-func TestFakeService_InterruptDispatch_RecordsDispatchInterruptedEvent(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	started := startAsyncByRequestID(t, service, "req-js-run-n-001")
-
-	result, err := service.InterruptDispatch(context.Background(), started.SessionID, InterruptDispatchRequest{
-		ControlRequest: ControlRequest{Reason: "stop bad run"},
-		DispatchID:     "disp-js-002",
-	})
-	if err != nil {
-		t.Fatalf("InterruptDispatch: %v", err)
-	}
-	if result.Outcome != LifecycleControlOutcomeAccepted {
-		t.Fatalf("outcome = %q, want ACCEPTED", result.Outcome)
-	}
-
-	dispatch, err := service.GetDispatch(context.Background(), started.SessionID, "disp-js-002")
-	if err != nil {
-		t.Fatalf("GetDispatch: %v", err)
-	}
-	if dispatch.Status != DispatchStatusInterrupted {
-		t.Fatalf("dispatch status = %q, want INTERRUPTED", dispatch.Status)
-	}
-	if dispatch.FailureDetail == nil || dispatch.FailureDetail.Message != "stop bad run" {
-		t.Fatalf("failureDetail = %#v, want stop bad run", dispatch.FailureDetail)
-	}
-
-	events, err := service.ReadEvents(context.Background(), started.SessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents: %v", err)
-	}
-	found := false
-	for _, raw := range events.Events {
-		var envelope canonicalFactoryEvent
-		if err := json.Unmarshal(raw, &envelope); err != nil {
-			t.Fatalf("unmarshal event: %v", err)
-		}
-		if envelope.Type != "DISPATCH_INTERRUPTED" {
-			continue
-		}
-		found = true
-		if envelope.Context.DispatchID == nil || *envelope.Context.DispatchID != "disp-js-002" {
-			t.Fatalf("dispatchId = %#v, want disp-js-002", envelope.Context.DispatchID)
-		}
-	}
-	if !found {
-		t.Fatal("DISPATCH_INTERRUPTED event missing from session events")
-	}
-
-	replayed, err := ReplayDispatchProjection(events.Events)
-	if err != nil {
-		t.Fatalf("ReplayDispatchProjection: %v", err)
-	}
-	if len(replayed) != 1 || replayed[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("replayed = %#v, want one interrupted dispatch", replayed)
-	}
-}
-
-func TestRestoreInterruptedDispatchResultSuppression_LateCompletionDoesNotReactivateRouting(t *testing.T) {
-	t.Parallel()
-	observedAt := time.Date(2026, 6, 20, 15, 0, 0, 0, time.UTC)
-	interrupted := DispatchSummary{
-		ID:     "disp-js-002",
-		Status: DispatchStatusInterrupted,
-		Phase:  "execute",
-		Label:  "audit",
-		FailureDetail: &DispatchFailureDetail{
-			Reason:  dispatchInterruptionFailureReasonCode,
-			Message: "operator stop",
-		},
-	}
-	state := &runtimeSessionState{
-		session: SessionReadResult{SessionID: "dur-sess-interrupt-late-001", Phase: "execute"},
-		dispatches: []DispatchSummary{
-			interrupted,
-		},
-		dispatchStatusTransitions: map[string][]DispatchStatus{
-			"disp-js-002": {DispatchStatusQueued, DispatchStatusRunning, DispatchStatusInterrupted},
-		},
-	}
-	preserved := snapshotInterruptedDispatches(state)
-
-	lateRecords := []factory.JavaScriptRuntimeRecord{{
-		Kind: factory.JavaScriptRecordKindChildDispatch,
-		ChildDispatch: &factory.JavaScriptChildDispatchRecord{
-			DispatchID:         "disp-js-002",
-			Status:             factory.JavaScriptChildDispatchStatusCompleted,
-			Label:              "audit",
-			ArtifactRef:        "artifact://child-artifact-late",
-			ProviderSessionRef: "provider-session-late",
-			Provider:           "fake",
-		},
-	}}
-	applyRuntimeExecutionRecordProjection(state, "dur-sess-interrupt-late-001", lateRecords, observedAt)
-	if state.dispatches[0].Status != DispatchStatusCompleted {
-		t.Fatalf("projected status = %q, want COMPLETED before suppression", state.dispatches[0].Status)
-	}
-	if state.session.Progress == nil || state.session.Progress.CompletedDispatches != 1 {
-		t.Fatalf("progress before suppression = %#v, want one completed dispatch", state.session.Progress)
-	}
-
-	restoreInterruptedDispatchResultSuppression(state, preserved)
-
-	if state.dispatches[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("status after suppression = %q, want INTERRUPTED", state.dispatches[0].Status)
-	}
-	if len(state.dispatches[0].OutputArtifactIDs) != 0 {
-		t.Fatalf("outputArtifactIds = %#v, want suppressed late output", state.dispatches[0].OutputArtifactIDs)
-	}
-	if len(state.dispatches[0].ProviderSessionRefs) != 1 || state.dispatches[0].ProviderSessionRefs[0].ID != "provider-session-late" {
-		t.Fatalf("providerSessionRefs = %#v, want late diagnostic preserved", state.dispatches[0].ProviderSessionRefs)
-	}
-	if state.session.Progress.CompletedDispatches != 0 {
-		t.Fatalf("completedDispatches = %d, want 0 after suppression", state.session.Progress.CompletedDispatches)
-	}
-	for _, artifact := range state.artifacts {
-		if artifact.DispatchID == "disp-js-002" && artifact.Kind == "CHILD_RESULT" {
-			t.Fatalf("artifact = %#v, want late child output suppressed", artifact)
-		}
-	}
-	transitions := state.dispatchStatusTransitions["disp-js-002"]
-	if len(transitions) != 3 || transitions[2] != DispatchStatusInterrupted {
-		t.Fatalf("statusTransitions = %#v, want queued/running/interrupted", transitions)
-	}
-}
-
-func TestApplyTerminalRuntimeProjection_PreservesInterruptedDispatchAndEvents(t *testing.T) {
-	t.Parallel()
-	observedAt := time.Date(2026, 6, 20, 15, 30, 0, 0, time.UTC)
-	sessionID := "dur-sess-interrupt-terminal-001"
-	startedAt := observedAt.Add(-time.Minute)
-	running := SessionReadResult{
-		SessionID: sessionID,
-		Status:    LifecycleStatusRunning,
-		Phase:     "execute",
-		Lifecycle: &LifecycleTimestamps{StartedAt: &startedAt},
-	}
-	events := AppendDispatchInterruptedEvent(
-		BuildCanonicalRuntimeSessionEvents(running, ResultReadResult{SessionID: sessionID}),
-		running,
-		DispatchSummary{ID: "disp-js-002", Status: DispatchStatusRunning, Phase: "execute"},
-		InterruptDispatchRequest{DispatchID: "disp-js-002", ControlRequest: ControlRequest{Reason: "operator stop"}},
-		DispatchStatusRunning,
-		canonicalEventSourceRuntimeService,
-		observedAt,
-	)
-	prior := &runtimeSessionState{
-		session: running,
-		dispatches: []DispatchSummary{{
-			ID:            "disp-js-002",
-			Status:        DispatchStatusInterrupted,
-			Phase:         "execute",
-			FailureDetail: dispatchInterruptionFailureDetail("operator stop"),
-		}},
-		dispatchStatusTransitions: map[string][]DispatchStatus{
-			"disp-js-002": {DispatchStatusQueued, DispatchStatusRunning, DispatchStatusInterrupted},
-		},
-		events: events,
-	}
-	lateRecords := []factory.JavaScriptRuntimeRecord{{
-		Kind: factory.JavaScriptRecordKindChildDispatch,
-		ChildDispatch: &factory.JavaScriptChildDispatchRecord{
-			DispatchID:  "disp-js-002",
-			Status:      factory.JavaScriptChildDispatchStatusCompleted,
-			Label:       "audit",
-			ArtifactRef: "artifact://child-artifact-late",
-		},
-	}}
-	terminal := runtimeSessionState{session: running}
-	applyRuntimeSuccessProjection(&terminal, sessionID, factory.JavaScriptRuntimeOutcome{
-		OK:      true,
-		Records: lateRecords,
-		Value:   factory.TypedValue{JSON: json.RawMessage("null")},
-	}, observedAt)
-
-	applyTerminalRuntimeProjection(prior, terminal, factory.JavaScriptRuntimeOutcome{
-		OK:      true,
-		Records: lateRecords,
-		Value:   factory.TypedValue{JSON: json.RawMessage("null")},
-	})
-
-	if prior.dispatches[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("dispatch status = %q, want INTERRUPTED", prior.dispatches[0].Status)
-	}
-	foundInterruptedEvent := false
-	for _, raw := range prior.events {
-		var envelope factoryEventEnvelope
-		if err := json.Unmarshal(raw, &envelope); err != nil {
-			continue
-		}
-		if envelope.Type == "DISPATCH_INTERRUPTED" {
-			foundInterruptedEvent = true
-			break
-		}
-	}
-	if !foundInterruptedEvent {
-		t.Fatal("DISPATCH_INTERRUPTED event missing after terminal projection")
-	}
-	if prior.session.Progress != nil && prior.session.Progress.CompletedDispatches != 0 {
-		t.Fatalf("completedDispatches = %d, want 0", prior.session.Progress.CompletedDispatches)
-	}
-	if prior.session.Status != LifecycleStatusInterrupted {
-		t.Fatalf("session status = %q, want INTERRUPTED", prior.session.Status)
-	}
-	if prior.session.ResultSummary == nil || prior.session.ResultSummary.ResultStatus != string(ResultStatusUnavailable) {
-		t.Fatalf("resultSummary = %#v, want UNAVAILABLE", prior.session.ResultSummary)
-	}
-	if prior.result.SessionStatus != LifecycleStatusInterrupted || prior.result.ResultStatus != ResultStatusUnavailable {
-		t.Fatalf("result = status %q session %q, want UNAVAILABLE/INTERRUPTED", prior.result.ResultStatus, prior.result.SessionStatus)
-	}
-}
-
-func TestReplaySessionProjection_PauseResumeLifecycleEventsDeriveStatus(t *testing.T) {
-	t.Parallel()
-	startedAt := time.Date(2026, 6, 11, 12, 0, 0, 0, time.UTC)
-	pausedAt := time.Date(2026, 6, 11, 12, 0, 5, 0, time.UTC)
-	resumedAt := time.Date(2026, 6, 11, 12, 0, 10, 0, time.UTC)
-	sessionID := "dur-sess-replay-pause-resume-001"
-
-	baseSession := SessionReadResult{
-		SessionID:        sessionID,
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: "JAVASCRIPT",
-		SourceHash:       "sha256:fixture",
-		Policy:           PolicyProjection{EffectiveHash: "sha256:policy"},
-		ResolvedSource:   ResolvedSource{SourceHash: "sha256:fixture"},
-		Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt},
-	}
-	baseResult := ResultReadResult{
-		SessionID:    sessionID,
-		ResultStatus: ResultStatusNotReady,
-	}
-
-	events := BuildCanonicalRuntimeSessionEvents(baseSession, baseResult)
-	events = AppendSessionLifecycleControlEvent(
-		events,
-		SessionReadResult{SessionID: sessionID, Status: LifecycleStatusPaused, OrchestratorKind: "JAVASCRIPT", Dialect: "you-workflow-v1"},
-		LifecycleStatusRunning,
-		LifecycleControlPause,
-		LifecycleControlOutcomeAccepted,
-		pausedAt,
-		canonicalEventSourceRuntimeService,
-		"",
-	)
-	events = AppendSessionLifecycleControlEvent(
-		events,
-		SessionReadResult{SessionID: sessionID, Status: LifecycleStatusRunning, OrchestratorKind: "JAVASCRIPT", Dialect: "you-workflow-v1"},
-		LifecycleStatusPaused,
-		LifecycleControlResume,
-		LifecycleControlOutcomeAccepted,
-		resumedAt,
-		canonicalEventSourceRuntimeService,
-		"",
-	)
-
-	session, result, err := ReplaySessionProjection(events)
-	if err != nil {
-		t.Fatalf("ReplaySessionProjection: %v", err)
-	}
-	if session.Status != LifecycleStatusRunning {
-		t.Fatalf("status = %q, want RUNNING", session.Status)
-	}
-	if result.SessionStatus != LifecycleStatusRunning {
-		t.Fatalf("result sessionStatus = %q, want RUNNING", result.SessionStatus)
-	}
-	if session.Lifecycle == nil || session.Lifecycle.PausedAt == nil || !session.Lifecycle.PausedAt.Equal(pausedAt) {
-		t.Fatalf("pausedAt = %#v, want %s", session.Lifecycle, pausedAt)
-	}
-	if session.Lifecycle.ResumedAt == nil || !session.Lifecycle.ResumedAt.Equal(resumedAt) {
-		t.Fatalf("resumedAt = %#v, want %s", session.Lifecycle.ResumedAt, resumedAt)
-	}
-
-	var lifecycleEnvelope canonicalFactoryEvent
-	if err := json.Unmarshal(events[2], &lifecycleEnvelope); err != nil {
-		t.Fatalf("unmarshal lifecycle event: %v", err)
-	}
-	if lifecycleEnvelope.Type != "SESSION_LIFECYCLE_CONTROL" {
-		t.Fatalf("event type = %q, want SESSION_LIFECYCLE_CONTROL", lifecycleEnvelope.Type)
-	}
-}
-
-func TestReplaySessionProjection_LegacyPauseResumeEventsDeriveStatus(t *testing.T) {
-	t.Parallel()
-	startedAt := time.Date(2026, 6, 11, 12, 0, 0, 0, time.UTC)
-	pausedAt := time.Date(2026, 6, 11, 12, 0, 5, 0, time.UTC)
-	resumedAt := time.Date(2026, 6, 11, 12, 0, 10, 0, time.UTC)
-	sessionID := "dur-sess-replay-legacy-pause-resume-001"
-	sessionSequence := 1
-	source := canonicalEventSourceRuntimeService
-	mustMarshalEvent := func(event canonicalFactoryEvent) json.RawMessage {
-		t.Helper()
-		raw, err := json.Marshal(event)
-		if err != nil {
-			t.Fatalf("json.Marshal event: %v", err)
-		}
-		return raw
-	}
-
-	baseSession := SessionReadResult{
-		SessionID:        sessionID,
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: "JAVASCRIPT",
-		SourceHash:       "sha256:fixture",
-		Policy:           PolicyProjection{EffectiveHash: "sha256:policy"},
-		ResolvedSource:   ResolvedSource{SourceHash: "sha256:fixture"},
-		Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt},
-	}
-	baseResult := ResultReadResult{
-		SessionID:    sessionID,
-		ResultStatus: ResultStatusNotReady,
-	}
-
-	events := BuildCanonicalRuntimeSessionEvents(baseSession, baseResult)
-	events = append(events,
-		mustMarshalEvent(canonicalFactoryEvent{
-			SchemaVersion: "v1alpha",
-			ID:            "event-session-paused",
-			Type:          "SESSION_PAUSED",
-			Context: canonicalFactoryEventContext{
-				Sequence:        3,
-				Tick:            3,
-				EventTime:       pausedAt,
-				SessionID:       &sessionID,
-				SessionSequence: &sessionSequence,
-				Source:          &source,
-			},
-			Payload: mustMarshalPayload(map[string]any{
-				"status":   string(LifecycleStatusPaused),
-				"pausedAt": pausedAt.Format(time.RFC3339),
-			}),
-		}),
-		mustMarshalEvent(canonicalFactoryEvent{
-			SchemaVersion: "v1alpha",
-			ID:            "event-session-resumed",
-			Type:          "SESSION_RESUMED",
-			Context: canonicalFactoryEventContext{
-				Sequence:        4,
-				Tick:            4,
-				EventTime:       resumedAt,
-				SessionID:       &sessionID,
-				SessionSequence: &sessionSequence,
-				Source:          &source,
-			},
-			Payload: mustMarshalPayload(map[string]any{
-				"status":    string(LifecycleStatusRunning),
-				"resumedAt": resumedAt.Format(time.RFC3339),
-			}),
-		}),
-	)
-
-	session, result, err := ReplaySessionProjection(events)
-	if err != nil {
-		t.Fatalf("ReplaySessionProjection: %v", err)
-	}
-	if session.Status != LifecycleStatusRunning {
-		t.Fatalf("status = %q, want RUNNING", session.Status)
-	}
-	if result.SessionStatus != LifecycleStatusRunning {
-		t.Fatalf("result sessionStatus = %q, want RUNNING", result.SessionStatus)
-	}
-	if session.Lifecycle == nil || session.Lifecycle.PausedAt == nil || !session.Lifecycle.PausedAt.Equal(pausedAt) {
-		t.Fatalf("pausedAt = %#v, want %s", session.Lifecycle, pausedAt)
-	}
-	if session.Lifecycle.ResumedAt == nil || !session.Lifecycle.ResumedAt.Equal(resumedAt) {
-		t.Fatalf("resumedAt = %#v, want %s", session.Lifecycle.ResumedAt, resumedAt)
-	}
-}
-
-// pkgmaintcheck:ignore-cyclomatic-complexity this regression keeps pause/resume append and no-op event immutability assertions together.
-func TestFakeService_PauseResumeAppendsLifecycleControlEventsWithoutNoOpMutation(t *testing.T) {
-	t.Parallel()
-	service, err := NewFakeServiceFromContractFixtures(contractFixturesPath(t), fakeServiceTestClock(), fileeffects.ContractFixtureReader(os.ReadFile))
-	if err != nil {
-		t.Fatalf("NewFakeServiceFromContractFixtures: %v", err)
-	}
-	started, err := service.StartAsync(context.Background(), StartRequest{
-		RequestID: "req-js-run-n-001",
-		Source: Source{
-			Kind:      factory.WorkflowSourceKindFactoryID,
-			FactoryID: "customer-support-triage",
-		},
-	})
-	if err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-
-	beforeEvents, err := service.ReadEvents(context.Background(), started.SessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents before pause: %v", err)
-	}
-	beforeCount := len(beforeEvents.Events)
-
-	paused, err := service.Pause(context.Background(), started.SessionID, ControlRequest{RequestID: "ctrl-pause-events-001"})
-	if err != nil {
-		t.Fatalf("Pause: %v", err)
-	}
-	if paused.Outcome != LifecycleControlOutcomeAccepted || paused.Status != LifecycleStatusPaused {
-		t.Fatalf("pause = %#v, want ACCEPTED/PAUSED", paused)
-	}
-
-	afterPause, err := service.ReadEvents(context.Background(), started.SessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents after pause: %v", err)
-	}
-	if len(afterPause.Events) != beforeCount+1 {
-		t.Fatalf("event count after pause = %d, want %d", len(afterPause.Events), beforeCount+1)
-	}
-	assertCanonicalEventEnvelope(t, afterPause.Events[len(afterPause.Events)-1], "SESSION_LIFECYCLE_CONTROL", "session-lifecycle-control/"+started.SessionID+"/2")
-
-	pauseNoOp, err := service.Pause(context.Background(), started.SessionID, ControlRequest{})
-	if err != nil {
-		t.Fatalf("Pause no-op: %v", err)
-	}
-	if pauseNoOp.Outcome != LifecycleControlOutcomeNoOp {
-		t.Fatalf("pause no-op outcome = %q, want NO_OP", pauseNoOp.Outcome)
-	}
-	afterNoOp, err := service.ReadEvents(context.Background(), started.SessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents after no-op: %v", err)
-	}
-	if len(afterNoOp.Events) != len(afterPause.Events) {
-		t.Fatalf("event count after no-op = %d, want unchanged %d", len(afterNoOp.Events), len(afterPause.Events))
-	}
-
-	resumed, err := service.Resume(context.Background(), started.SessionID, ControlRequest{RequestID: "ctrl-resume-events-001"})
-	if err != nil {
-		t.Fatalf("Resume: %v", err)
-	}
-	if resumed.Outcome != LifecycleControlOutcomeAccepted || resumed.Status != LifecycleStatusRunning {
-		t.Fatalf("resume = %#v, want ACCEPTED/RUNNING", resumed)
-	}
-
-	afterResume, err := service.ReadEvents(context.Background(), started.SessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents after resume: %v", err)
-	}
-	if len(afterResume.Events) != len(afterPause.Events)+1 {
-		t.Fatalf("event count after resume = %d, want %d", len(afterResume.Events), len(afterPause.Events)+1)
-	}
-
-	replayed, _, err := ReplaySessionProjection(afterResume.Events)
-	if err != nil {
-		t.Fatalf("ReplaySessionProjection: %v", err)
-	}
-	if replayed.Status != LifecycleStatusRunning {
-		t.Fatalf("replayed status = %q, want RUNNING", replayed.Status)
 	}
 }
